requirements.txt
attrs==22.1.0
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
kiwisolver==1.4.4
matplotlib==3.6.1
networkx==2.8.7
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
PyGuitarPro==0.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
